Move do_set_price() and do_buy_item() to buy_sell.rs
