+++ frame/nfts/src/functions.rs
@@ -18,10 +18,7 @@
 //! Various pieces of common functionality.
 
 use super::*;
-use frame_support::{
-	ensure,
-	traits::{ExistenceRequirement, Get},
-};
+use frame_support::{ensure, traits::Get};
 use sp_runtime::{DispatchError, DispatchResult};
 
 impl<T: Config<I>, I: 'static> Pallet<T, I> {
@@ -274,94 +271,6 @@ impl<T: Config<I>, I: 'static> Pallet<T, I> {
 		}
 
 		Self::deposit_event(Event::Burned { collection, item, owner });
-		Ok(())
-	}
-
-	pub fn do_set_price(
-		collection: T::CollectionId,
-		item: T::ItemId,
-		sender: T::AccountId,
-		price: Option<ItemPrice<T, I>>,
-		whitelisted_buyer: Option<T::AccountId>,
-	) -> DispatchResult {
-		ensure!(
-			Self::is_pallet_feature_enabled(PalletFeature::Trading),
-			Error::<T, I>::MethodDisabled
-		);
-
-		let details = Item::<T, I>::get(&collection, &item).ok_or(Error::<T, I>::UnknownItem)?;
-		ensure!(details.owner == sender, Error::<T, I>::NoPermission);
-
-		let collection_config = Self::get_collection_config(&collection)?;
-		ensure!(
-			collection_config.is_setting_enabled(CollectionSetting::TransferableItems),
-			Error::<T, I>::ItemsNonTransferable
-		);
-
-		let item_config = Self::get_item_config(&collection, &item)?;
-		ensure!(
-			item_config.is_setting_enabled(ItemSetting::Transferable),
-			Error::<T, I>::ItemLocked
-		);
-
-		if let Some(ref price) = price {
-			ItemPriceOf::<T, I>::insert(&collection, &item, (price, whitelisted_buyer.clone()));
-			Self::deposit_event(Event::ItemPriceSet {
-				collection,
-				item,
-				price: *price,
-				whitelisted_buyer,
-			});
-		} else {
-			ItemPriceOf::<T, I>::remove(&collection, &item);
-			Self::deposit_event(Event::ItemPriceRemoved { collection, item });
-		}
-
-		Ok(())
-	}
-
-	pub fn do_buy_item(
-		collection: T::CollectionId,
-		item: T::ItemId,
-		buyer: T::AccountId,
-		bid_price: ItemPrice<T, I>,
-	) -> DispatchResult {
-		ensure!(
-			Self::is_pallet_feature_enabled(PalletFeature::Trading),
-			Error::<T, I>::MethodDisabled
-		);
-
-		let details = Item::<T, I>::get(&collection, &item).ok_or(Error::<T, I>::UnknownItem)?;
-		ensure!(details.owner != buyer, Error::<T, I>::NoPermission);
-
-		let price_info =
-			ItemPriceOf::<T, I>::get(&collection, &item).ok_or(Error::<T, I>::NotForSale)?;
-
-		ensure!(bid_price >= price_info.0, Error::<T, I>::BidTooLow);
-
-		if let Some(only_buyer) = price_info.1 {
-			ensure!(only_buyer == buyer, Error::<T, I>::NoPermission);
-		}
-
-		T::Currency::transfer(
-			&buyer,
-			&details.owner,
-			price_info.0,
-			ExistenceRequirement::KeepAlive,
-		)?;
-
-		let old_owner = details.owner.clone();
-
-		Self::do_transfer(collection, item, buyer.clone(), |_, _| Ok(()))?;
-
-		Self::deposit_event(Event::ItemBought {
-			collection,
-			item,
-			price: price_info.0,
-			seller: old_owner,
-			buyer,
-		});
-
 		Ok(())
 	}
